+++ messages/pitchrollheading.go
@@ -18,18 +18,18 @@
 package messages
 
 type PitchRollHeadingMessage struct {
-	Pitch           float32
-	Roll            float32
-	HeadingTrue     float32
-	HeadingMagnetic float32
+	Pitch           float64
+	Roll            float64
+	HeadingTrue     float64
+	HeadingMagnetic float64
 }
 
 func NewPitchRollHeadingMessage(data []float32) PitchRollHeadingMessage {
 	return PitchRollHeadingMessage{
-		Pitch:           data[0],
-		Roll:            data[1],
-		HeadingTrue:     data[2],
-		HeadingMagnetic: data[3],
+		Pitch:           float64(data[0]),
+		Roll:            float64(data[1]),
+		HeadingTrue:     float64(data[2]),
+		HeadingMagnetic: float64(data[3]),
 	}
 }
